requirements.txt
numpy
pyrtlsdr